feat/examples: reimplement client_stress_test example
