+++ safe_core/examples/email.rs
