+++ safe_core/examples/email_stress_test.rs
