+++ safe_core/examples/self_authentication.rs
